Support waveshare35b-v2

+++ system_backup.sh
@@ -51,8 +51,12 @@ fi
 if [[ "$deb_ver" < "12.1" ]]; then
 sudo cp -rf ./usr/99-fbturbo.conf-original /usr/share/X11/xorg.conf.d/99-fbturbo.conf
 fi
+if [ -f /etc/rc.local ]; then
 sudo cp -rf /etc/rc.local ./.system_backup/
 sudo cp -rf ./etc/rc.local-original /etc/rc.local
+else
+sudo cp ./etc/rc.local-original /etc/rc.local
+fi
 
 sudo cp -rf /etc/modules ./.system_backup/
 sudo cp -rf ./etc/modules-original /etc/modules
